requirements.txt
streamlit==1.44.1
anthropic==0.51.0
edge-tts==6.1.19
pydub==0.25.1
pdfplumber==0.11.6
python-docx==1.1.2
python-dotenv==1.0.1
